Alle notwendigen Testklassen hinzugefügt
