+++ MEGAtroniker_test/grind/core/impl/SpielmodellMEGAtronikerTest.java
@@ -1,0 +1,21 @@
+package grind.core.impl;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class SpielmodellMEGAtronikerTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+
+    @Test
+    public void addMonster() {
+    }
+
+    @Test
+    public void removeMovable() {
+    }
+}
