+++ MEGAtroniker_test/grind/core/impl/SpielsteuerungMEGAtronikerTest.java
@@ -1,0 +1,25 @@
+package grind.core.impl;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class SpielsteuerungMEGAtronikerTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+
+    @Test
+    public void pruefeKollisionen() {
+    }
+
+    @Test
+    public void pruefeUmgebung() {
+    }
+
+    @Test
+    public void setImages() {
+    }
+}
