+++ MEGAtroniker_test/grind/movables/monster/DornPflanzeMEGAtronikerTest.java
@@ -1,0 +1,41 @@
+package grind.movables.monster;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class DornPflanzeMEGAtronikerTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+
+    @Test
+    public void bewege() {
+    }
+
+    @Test
+    public void getGroesse() {
+    }
+
+    @Test
+    public void vorBetreten() {
+    }
+
+    @Test
+    public void getSpielmodell() {
+    }
+
+    @Test
+    public void setSpielmodell() {
+    }
+
+    @Test
+    public void setGeschwindigkeit() {
+    }
+
+    @Test
+    public void getGeschwindigkeit() {
+    }
+}
